==> notebooks/02_SDGVariables.ipynb <==
# Load required libraries

# Import Data Commons

import pandas as pd
import numpy as np
import datacommons_pandas as dc

# Import other required libraries
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import pandas as pd

import json
import time

import ast

import os
from dotenv import load_dotenv

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"
from IPython.display import clear_output, display
from IPython.core.display import HTML, JSON

import requests
from bs4 import BeautifulSoup


import sys
sys.path.append('../') # add the project's root directory to the system path

from utils_excel import * # import the utils_excel module from the project's root directory


# Load key

load_dotenv()
api_key = os.getenv("DC_KEY")

---

### SPARQL Query

# set up the API endpoint URL
url = "https://api.datacommons.org/v1/query"
headers = {"X-API-Key": api_key}

### REST API

def call_api(endpoint, parameters):
    url = f"http://api.datacommons.org/{endpoint}{parameters}?key={api_key}"
    print(f"http://api.datacommons.org/{endpoint}{parameters}?key={api_key}")
    response = requests.get(url)
    return json.loads(response.content)

## 1. Get list of all `SDG_Series`

query = """
        SELECT DISTINCT ?dcid ?name  
        WHERE {
          ?x typeOf SDG_Series .
          ?x dcid ?dcid .
          ?x name ?name .
        }
        ORDER BY DESC(?dcid)
        """
request_body = {"sparql": query}

# send the request to the API endpoint
response = requests.post(url, headers=headers, data=json.dumps(request_body))

results = json.loads(response.content)['rows']

df_SDG_SERIES = pd.DataFrame([('SDG_Series', r['cells'][0]['value'], r['cells'][1]['value'], f"https://datacommons.org/browser/{r['cells'][0]['value']}") for r in results], columns=['typeOf', 'dcid', 'name', 'DC Graph Browser URL'])
df_SDG_SERIES

write_to_excel(df_SDG_SERIES, '../data/output/SDG_Series.xlsx', 'SDG_Series', 90)

## 2. Find the statistical variables associated with each SDG_Series

statistical_variables = []
statistical_variables_log = []

endpoint = 'v1/property/values'

for index, sdg_series_id in df_SDG_SERIES['dcid'].items():
    
    clear_output(wait=True)
    
    parameters = f'/in/{sdg_series_id}/populationType'
    
    respons_data = call_api(endpoint, parameters)
    try:
        x = pd.DataFrame(respons_data['values'])
        x['SDG_Series'] = sdg_series_id
        statistical_variables.append(x)
        log_entry = {'SDG_Series': sdg_series_id, 'No_variables': len(x)}
    except:
        log_entry = {'SDG_Series': sdg_series_id, 'No_variables': 0}
    print(log_entry)
    statistical_variables_log.append(log_entry)
    clear_output(wait=True)
statistical_variables = pd.concat(statistical_variables)
statistical_variables_log =pd.DataFrame(statistical_variables_log)

write_to_excel(statistical_variables, '../data/output/statistical_variables.xlsx', 'StatisticalVariables', 90)
write_to_excel(statistical_variables_log, '../data/output/statistical_variables_log.xlsx', 'StatisticalVariables', 90)

## List the properties of each statistical variable

variable_properties = []

for i, v in statistical_variables['dcid'].items():

    # if v!= 'sdg/VC_VOV_GDSD_AGE_FEMALE':
    #     continue
        
    clear_output(wait=True)
    
    d = dict()
    d['dcid'] = v
    
    properties = call_api('v1/properties', f"/out/{v}")['properties']
    print('\n')
    for p in properties:
        property_values = call_api('v1/property/values', f'/out/{v}/{p}')['values']
        if p != 'name':
            if len(property_values)>1:
                d[p] = [item['dcid'] for item in property_values]
                d[f"{p}__name"] = [item['name'] for item in property_values]
            else:
                d[p] = property_values[0]['dcid'] 
                d[f"{p}__name"] = property_values[0]['name']
        else: 
            d[p] = property_values[0]['value'] 
            
    
    variable_properties.append(d)
variable_properties = pd.DataFrame(variable_properties)
variable_properties = variable_properties.fillna('',)    
variable_properties = variable_properties.astype(str) 
variable_properties = variable_properties.drop_duplicates()
variable_properties



write_to_excel(variable_properties, '../data/output/Variable_Properties.xlsx', 'Variable_Properties', 90)

list(variable_properties)
