==> notebooks/04_SDGEnumerations.ipynb <==
# Explore SDG Enumerations


For all SDG statistical variables currently in DC, we have the list of ‘constraintProperties’, i.e., the list of dimensions that are actively used by at least one statistical variable.  Each of these properties has a set of values, which in turn as of a specific Enumeration class.  

# Import Data Commons

import pandas as pd
import numpy as np
import datacommons_pandas as dc

# Import other required libraries
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import pandas as pd

import json
import time

import ast

import os
from dotenv import load_dotenv

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"
from IPython.display import clear_output, display
from IPython.core.display import HTML, JSON

import requests
from bs4 import BeautifulSoup


import sys
sys.path.append('../') # add the project's root directory to the system path

from utils_excel import * # import the utils_excel module from the project's root directory


---

load_dotenv()
api_key = os.getenv("DC_KEY")

### SPARQL Query

# set up the API endpoint URL
url = "https://api.datacommons.org/v1/query"
headers = {"X-API-Key": api_key}

### REST API

def call_api(endpoint, parameters):
    url = f"http://api.datacommons.org/{endpoint}{parameters}?key={api_key}"
    print(f"http://api.datacommons.org/{endpoint}{parameters}")
    response = requests.get(url)
    return json.loads(response.content)

---

variable_properties = pd.read_excel('../data/output/Variable_Properties.xlsx').fillna('')
variable_properties.head(3)

### 1. Obtain properties directly linked to SDG statistical variables through the `onstraintProperties` relation

ConstraintProperties = list(variable_properties['constraintProperties'])


ConstraintProperties_unique = []

for element in ConstraintProperties:
    # Check if the element is a simple string
    if isinstance(element, str) and not element.startswith("["):
        if element not in ConstraintProperties_unique:
            ConstraintProperties_unique.append(element)
    # Check if the element is a complex string that can be parsed as an array of simple strings
    elif isinstance(element, str) and element.startswith("[") and element.endswith("]"):
        parsed_array = ast.literal_eval(element)
        for value in parsed_array:
            if value not in ConstraintProperties_unique:
                ConstraintProperties_unique.append(value)

ConstraintProperties_unique = sorted(ConstraintProperties_unique)
ConstraintProperties_unique.remove('')
print(f"{len(ConstraintProperties_unique)=}")
ConstraintProperties_unique

enumerations = []

for idx, dimension in enumerate(ConstraintProperties_unique):

    dim_values = sorted(list(set([x for x in list(variable_properties[dimension]) if x != ''])))
    
    for dv in dim_values:
        row = dict()
        row['dimension'] = dimension
        row['dim_value'] = dv
        x = call_api('v1/property/values', f'/out/{dv}/typeOf')['values']
        row['dim_enumeration'] =  x[0]['dcid']
        row['dim_enumeration_name'] =  x[0]['name']
        enumerations.append(row)
    
enumerations = pd.DataFrame(enumerations)
enumerations
    
    
    

Dimension_Enumeration = enumerations[['dimension', 'dim_enumeration', 'dim_enumeration_name']].drop_duplicates()

new_col_names = {'dimension': 'dimension', 'dim_enumeration': 'enumeration', 'dim_enumeration_name': 'enumeration_name'}
Dimension_Enumeration = Dimension_Enumeration.rename(columns=new_col_names)
Dimension_Enumeration = Dimension_Enumeration.reset_index(drop=True)
Dimension_Enumeration

write_to_excel(Dimension_Enumeration, '../data/output/Dimension_Enumeration.xlsx', 'Dimension_Enumeration', 90)

for idx, e in enumerate(Dimension_Enumeration['enumeration']):
    code_list = call_api('v1/property/values', f'/in/{e}/typeOf')['values']
    write_to_excel(pd.DataFrame(code_list), f'../data/output/Enumerations/{e}.xlsx', 'Enumeration', 90)
